Enable Priority Flow Control for interfaces
Enable Priority Flow Contorl for all discovered interfaces when in L3
mode and supplied with '--pfc 0,1,2,3,4,5,6,7' or a comma separated list
priorities (0-7). For this lldptool binary is looked for via the PATH
environment variable. lldpad needs to be runnig in an adjacent container
in the same pod in order for lldptool to send commands via the AF_LOCAL
socket to lldpad in the same network namespace.

Fixes by tuomas.katila@intel.com.

Signed-off-by: Patrik Flykt <patrik.flykt@intel.com>