Find libraries to link statically
When linking statically, all libraries have to be specified separately
except for libc. libpcap has different dependencies in Debian and Ubuntu,
possibly in other distros as well. In Ubuntu the library includes a
dependency on libibverbs, which has additional dependencies of its own.

Use pkg-confg to find the correct set of dependencies in the OS where
binaries are being built.

In addition, the host system may use a different OS from the one used in
building images, thus the dependencies are resolved independently of the
host OS when building the image.

Signed-off-by: Patrik Flykt <patrik.flykt@intel.com>